refactor(rag): split pipeline.ts + run vector and BM25 in parallel (Plan B+C)
Sprint 3 collapses the two near-identical retrieval blocks (runPipeline +
runRetrieval) into a single runRetrievalCore so ask() and askStream()
share retrieval semantics. Plan C dispatches vector and BM25 inside one
Promise.all — the historical "BM25 first" cache-locality argument no
longer applies (the worker pool owns its own SQLite handles and does not
share a page cache with the main thread).

Splits:
- analyzer.ts (393 lines) — query intent + AnalyzedQuery + helpers
  (numbersToDigits, isModifierNorm, normalizePeriodicTitle, isSectoralNorm,
  isFundamentalRank, describeNormScope, resolveNormsByName)
- bm25-dispatch.ts (280 lines) — 5-stage fan-out, closes the runBm25/
  runBm252 duplication flagged by the PR #34 reviewer; preserves per-
  stage sync fallback resilience and the saturation counter
- retrieval.ts (~700 lines) — runRetrievalCore + computeBoosts +
  applyLegalHierarchyBoost + articleTypePenalty + getArticleData
- synthesis.ts (~500 lines) — prompts, evidence builder, JSON +
  streaming synth, citation verification, citizen-summary backfill
- pipeline.ts shrinks 3241 → 727 lines and is now an orchestrator
  (RagPipeline class + AskRequest/AskResponse types + Citation re-export)

Public surface preserved:
- RagPipeline class (used by routes/ask.ts)
- articleTypePenalty / normalizePeriodicTitle (used by tests)
- AskRequest / AskResponse / Citation types

New eval gate:
- packages/api/research/eval-gate.ts replays the 65-question omnibus set
  through a new RagPipeline._retrieveForEval hook and computes
  R@1 / R@5 / R@10 (norm-level). Used to assert the refactor is
  observably identical to pre-refactor production behaviour.
